@@ -6,8 +6,8 @@ import org.pircbotx.output.OutputRaw;
 
 public class CustomBotFactory extends BotFactory {
 	
-	private double messageLimit;
-	private double messageLimitInterval;
+	private final double messageLimit;
+	private final double messageLimitInterval;
 	
 	public CustomBotFactory(double messageLimit, double messageLimitInterval) {
 		this.messageLimit = messageLimit;
